backend/pyproject.toml
[project]
name = "hireflow-backend"
version = "1.0.0"
description = "HireFlow AI - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "PyMuPDF>=1.23.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "openai>=1.6.0",
    "google-generativeai>=0.3.0",
    "python-docx>=1.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
]
